Added parent file check to other functions
